server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==8.4.0
pip==25.0
protobuf==4.25.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.45.3
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
wheel==0.45.1
win_inet_pton==1.1.0
